backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
asttokens==3.0.0
attrs==23.2.0
autopep8==2.3.1
beautifulsoup4==4.13.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-appconf==1.0.5
django-cors-headers==4.2.0
django-extensions==3.2.3
django-imagekit==4.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fonttools==4.55.3
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jiter==0.10.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.82.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pilkit==2.0
Pillow==8.3.2
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2023.3
pywin32==310
PyYAML==6.0.1
pyzmq==26.3.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
selenium==4.30.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.21.0
